==> 01_data-collection.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.remote.webelement import WebElement
from bs4 import BeautifulSoup
from selenium.webdriver.common.keys import Keys

driver = webdriver.Chrome()
driver.get('https://cjs.shelbycountytn.gov/CJS')
driver.find_element(By.ID, "dropdownMenu1").click()

WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, "//a[@href = '/CJS/Account/Login']"))).click()

driver.find_element(By.ID,"UserName").send_keys('golderlawfirm@gmail.com')
driver.find_element(By.ID,"Password").send_keys('5Ram0n9!')

driver.find_element(By.XPATH, "//button").click()

driver.find_element(By.XPATH, "//a[@href = '/CJS/Home/Dashboard/29']").click()

# Run Range 10 through 50 or so

for i in range(250):
    court = []
    defendant = []
    charges = []
    bond = []
    nums = 21000900 + 100*i
    for x in range(nums,nums+100):
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(Keys.CONTROL, 'a')
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(Keys.BACKSPACE)
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(str(x))
        driver.find_element(By.ID, "btnSSSubmit").click()
        try:
            WebDriverWait(driver, 5).until(EC.element_to_be_clickable((By.LINK_TEXT, str(x)))).click()
        except:
            pass
        time.sleep(2)
        try:
            court.append(driver.find_element(By.ID, "divCaseInformation_body").text)
            defendant.append(driver.find_element(By.ID, "divPartyInformation_body").text)
            charges.append(driver.find_element(By.XPATH, "//div[contains(@id, 'ChargeGrid')]").text)
            bond.append(driver.find_element(By.ID, "BondsGrid").text.split('$')[1])
        except:
            pass
        driver.back()
    df = pd.DataFrame(data= zip(court,defendant,charges,bond), columns = ['court','defendant','charges','bond'])
    df.to_csv(f'{nums}_{nums+100}.csv',index=False)

# Run Range 10 through 50 or so

for i in range(250):
    court = []
    defendant = []
    charges = []
    bond = []
    nums = 19000000 + 100*i
    for x in range(nums,nums+100):
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(Keys.CONTROL, 'a')
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(Keys.BACKSPACE)
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(str(x))
        driver.find_element(By.ID, "btnSSSubmit").click()
        try:
            WebDriverWait(driver, 5).until(EC.element_to_be_clickable((By.LINK_TEXT, str(x)))).click()
        except:
            pass
        time.sleep(2)
        try:
            court.append(driver.find_element(By.ID, "divCaseInformation_body").text)
            defendant.append(driver.find_element(By.ID, "divPartyInformation_body").text)
            charges.append(driver.find_element(By.XPATH, "//div[contains(@id, 'ChargeGrid')]").text)
            bond.append(driver.find_element(By.ID, "BondsGrid").text.split('$')[1])
        except:
            pass
        driver.back()
    df = pd.DataFrame(data= zip(court,defendant,charges,bond), columns = ['court','defendant','charges','bond'])
    df.to_csv(f'{nums}_{nums+100}.csv',index=False)

# Run Range 10 through 50 or so

for i in range(250):
    court = []
    defendant = []
    charges = []
    bond = []
    nums = 20000000 + 100*i
    for x in range(nums,nums+100):
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(Keys.CONTROL, 'a')
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(Keys.BACKSPACE)
        driver.find_element(By.CLASS_NAME, "form-control").send_keys(str(x))
        driver.find_element(By.ID, "btnSSSubmit").click()
        try:
            WebDriverWait(driver, 5).until(EC.element_to_be_clickable((By.LINK_TEXT, str(x)))).click()
        except:
            pass
        time.sleep(2)
        try:
            court.append(driver.find_element(By.ID, "divCaseInformation_body").text)
            defendant.append(driver.find_element(By.ID, "divPartyInformation_body").text)
            charges.append(driver.find_element(By.XPATH, "//div[contains(@id, 'ChargeGrid')]").text)
            bond.append(driver.find_element(By.ID, "BondsGrid").text.split('$')[1])
        except:
            pass
        driver.back()
    df = pd.DataFrame(data= zip(court,defendant,charges,bond), columns = ['court','defendant','charges','bond'])
    df.to_csv(f'{nums}_{nums+100}.csv',index=False)
