==> 02_data-cleaning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()
pd.set_option('display.max_columns',150)

import re
import os
import glob

os.path.join('./outputs/','*.csv')

files = glob.glob(os.path.join('./outputs/','*.csv'))

cases = pd.concat(map(pd.read_csv,files),axis = 0)

cases.head()

cases.info()

pd.set_option('display.max_colwidth',200)

for col in cases:
    cases[col] = cases[col].str.replace('\r','')

cases['cl'] = [set(re.findall(r'Misdemeanor\s\D|Felony\s\D',charge)) for charge in cases['charges']]

cases['hom'] = cases['charges'].apply(lambda x: 1 if 'murder' in x.lower() else 0)

cases['gun'] = cases['charges'].apply(lambda x: 1 if 'firearm' in x.lower() else 1 if 'handgun' in x.lower() else 0)

cases['rape'] = cases['charges'].apply(lambda x: 1 if 'rape' in x.lower() else 0)

cases['sex'] = cases['charges'].apply(lambda x: 1 if 'sexual' in x.lower() else 0)

cases['robb'] = cases['charges'].apply(lambda x: 1 if 'robbery' in x.lower() else 0)

cases['burg'] = cases['charges'].apply(lambda x: 1 if 'burg' in x.lower() else 0)

cases['drug'] = cases['charges'].apply(lambda x: 1 if 'upcs' in x.lower()\ 
                                       else 1 if 'controlled' in x.lower()\
                                       else 1 if 'meth' in x.lower()\
                                       else 1 if 'drug' in x.lower() else 0)

cases['driv'] = cases['charges'].apply(lambda x: 1 if 'driving' in x.lower() else 0)

cases['ch'] = [set(re.findall('\d{2}\-\d{2}-\d{3,6}',charge)) for charge in cases['charges']]

cases['name'] = cases['defendant'].str.split('\n',expand=True)[1]
cases['dob'] = cases['defendant'].str.split('\n',expand=True)[3]
cases['race'] = cases['defendant'].str.split('\n',expand=True)[5]
cases['zip'] = cases['defendant'].str.split('\n',expand=True)[8][-5:]
cases['atty'] = cases['defendant'].str.split('\n',expand=True)[11]
cases['retained'] = cases['defendant'].str.split('\n',expand=True)[12]

cases['case_no'] = cases['court'].str.split('\n',expand = True)[5]
cases['judge'] = cases['court'].str.split('\n',expand = True)[9]
cases['date'] = cases['court'].str.split('\n',expand = True)[11]
cases['type'] = cases['court'].str.split('\n',expand = True)[13]
cases['status'] = cases['court'].str.split('\n',expand = True)[15]
cases['bond_amt'] = cases['bond'].str.split(' ', n=1, expand=True)[0]
cases['bond_status'] = cases['bond'].str.split(' ', n=1, expand=True)[1]

cases = cases[cases['cl'] != set()]

cases = cases.iloc[:,4:]

cases['bond_amt'] = cases['bond_amt'].str.replace(',','')

cases['bond_amt'] = pd.to_numeric(cases['bond_amt'])

cases

cases['cl'].value_counts().head(50)

cases['dob'] = pd.to_datetime(cases['dob'], errors = 'coerce')

cases['dob'].value_counts()

cases['race'].value_counts()

cases['race'] = cases['race'].apply(lambda x: x if x in ['Black','White','Asian'] else 'Other')

cases['zip'].value_counts()

cases['atty'] = cases['atty'].str.upper()

cases['atty'].value_counts()

cases['retained'].value_counts()

cases['atty'] = cases.apply(lambda x: x['retained'] if x['retained'] not in ['Retained','Court Appointed'] else x['atty'], axis = 1)

cases['case_no'].sort_values()

cases['atty'] = cases.apply(lambda x: x['retained'] if x['retained'] not in ['Retained','Court Appointed'] else x['atty'], axis = 1)

cases['date'] = cases.apply(lambda x: x['judge'] if x['date'] == 'State Warrant' else x['date'], axis = 1)

cases['judge'] = cases.apply(lambda x: x['judge'] if x['judge'] in ['Gilbert, Greg','Lucchesi, Ronald','Anderson, William Bill',\
                                                                 'Wilson, Lee','Massey, Karen','Renfroe, Sheila B.','Montesi, Louis J., Jr.',\
                                                                 'Johnson, Christian R.','Turner, William C.','Dwyer, Tim','Lambert-Ryan, Loyce'\
                                                                 'Skahan, Gerald'] else 'Other',axis = 1)

cases['type'].value_counts()

cases['status'].value_counts()

cases['bond_amt'].value_counts().head(50)

cases['bond_status'].value_counts().head(50)

cases.isnull().sum().head(50)

cases['ch'].value_counts().head(50)

cases.loc[cases['date'] >= '02/15/2023','era'] = 2

cases.loc[(cases['date'] >= '01/01/2023') & (cases['date'] < '02/15/2023'),'era'] = 1

cases.loc[(cases['date'] >= '09/01/2022') & (cases['date'] <= '12/31/2022'),'era'] = 1

cases.loc[cases['date'] < '09/01/2022','era'] = 0

cases['log_bond'] = np.log(cases['bond_amt'] + 1)

cases.to_csv('ody.csv',index=False)
