==> 04_regression-model.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LinearRegression, Lasso, ElasticNet, Ridge

from sklearn.neighbors import KNeighborsRegressor

from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor, AdaBoostRegressor

from sklearn.preprocessing import OneHotEncoder, StandardScaler

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()

from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.linear_model import LinearRegression, ElasticNet, LogisticRegression
from sklearn.neighbors import KNeighborsClassifier, KNeighborsRegressor
from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor,AdaBoostClassifier, AdaBoostRegressor, BaggingRegressor, BaggingClassifier
from sklearn.svm import SVR, SVC
import xgboost as kg

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.metrics import mean_squared_error, f1_score

cases = pd.read_csv('bonds.csv')

cases.info()

cases

X = cases[['ch','cl','judge','race']]
y = cases['log_bond']

ohe = OneHotEncoder()
X = ohe.fit_transform(X)

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

base = [np.mean(y_train)]*len(y_test)

mean_squared_error(y_test,base,squared=False)

knn = KNeighborsRegressor(n_neighbors=200)
knn.fit(X_train,y_train)
mean_squared_error(y_train,knn.predict(X_train),squared = False),mean_squared_error(y_test,knn.predict(X_test),squared = False)

dt = DecisionTreeRegressor()
dt.fit(X_train,y_train)
mean_squared_error(y_test,dt.predict(X_test),squared=False)

gs = GridSearchCV(knn,
                  param_grid = {'n_neighbors':[3,5,7,10]},
                  cv=5)

gs.fit(X_train,y_train)
gs.best_estimator_

knn = KNeighborsRegressor(n_neighbors = 10)
knn.fit(X_train,y_train)
mean_squared_error(y_train,knn.predict(X_train),squared = False),mean_squared_error(y_test,knn.predict(X_test),squared = False)

model_dict = {'knn': 
              {'model': 
               ('knn', KNeighborsRegressor()),
               'params':
               {"knn__n_neighbors": [3,5,7,10],
                "knn__p":[1,2]}},
             'lr': 
              {'model': 
               ('lr', LinearRegression()),
               'params':
               {"lr__fit_intercept": [True,False]}},
             'abc': 
              {'model': 
               ('abc', AdaBoostRegressor()),
               'params':
               {"abc__n_estimators": [30,50,100],
                "abc__learning_rate": [0.2,0.4,0.6]}}
              }

for key, value in model_dict.items():
    print(value['model'][1],value['params'])


best_models = {}

for key, value in model_dict.items():
    gs = GridSearchCV(value['model'][0], # set as iteration of pipe
                  param_grid = value['params'], # set as iteration of pipe_params
                      cv=5, # Cross Validation of 5
                      n_jobs=-1) # Unlock CPU to help process

    gs.fit(X_train, y_train)

    best_models[key] = {'model': gs.best_estimator_,
                        'params': gs.best_params_}
    print(best_models)

for key, value in best_models.items():
    value['model'].fit(X_train,y_train)
    print(value['model'].score(X_train, y_train), value['model'].score(X_test, y_test))

base = [np.mean(y_train)] * len(y_test)
mean_squared_error(y_test,base, squared = False)

resids = lr.predict(X_test) - y_test

resids

lr = LinearRegression()
lr.fit(X_train,y_train)
mean_squared_error(y_test,lr.predict(X_test),squared = False)

knn = KNeighborsRegressor(n_neighbors = 50)
knn.fit(X_train,y_train)
mean_squared_error(y_test, knn.predict(X_test), squared = False)

dt = DecisionTreeRegressor()
dt.fit(X_train,y_train)
mean_squared_error(y_test, dt.predict(X_test), squared = False)

pd.DataFrame(lr.predict(X_test),y_test,columns = ['pred','true'])
