==> 05_Classification.ipynb <==
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LinearRegression, Lasso, ElasticNet, Ridge

from sklearn.preprocessing import OneHotEncoder, StandardScaler

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()

from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.linear_model import LinearRegression, ElasticNet, LogisticRegression
from sklearn.neighbors import KNeighborsClassifier, KNeighborsRegressor
from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor,AdaBoostClassifier, AdaBoostRegressor, BaggingRegressor, BaggingClassifier
from sklearn.svm import SVR, SVC
import xgboost as kg

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.metrics import mean_squared_error, f1_score

from sklearn.naive_bayes import GaussianNB, MultinomialNB, BernoulliNB

from sklearn.metrics import confusion_matrix

cases = pd.read_csv('ody.csv')

cases.head()

cases['ror'] = cases['bond_amt'].apply(lambda x: 1 if x == 0 else 0)

cases.isnull().sum().head(50)

X = cases[['ch','race']]

y = cases['ror']

ohe = OneHotEncoder()
X = ohe.fit_transform(X).toarray()

X = pd.DataFrame(X, columns=ohe.get_feature_names_out())

X_train,X_test,y_train,y_test = train_test_split(X, y, stratify=y, random_state=42)

X_train.shape,X_test.shape,y_train.shape,y_test.shape

base = [0]*len(y_test)

confusion_matrix(y_test,base)

logr = LogisticRegression(C=1000)
logr.fit(X_train,y_train)
confusion_matrix(y_test,logr.predict(X_test))

knn = KNeighborsClassifier(n_neighbors = 5)
knn.fit(X_train,y_train)
confusion_matrix(y_test,knn.predict(X_test))

gnb = GaussianNB()
gnb.fit(X_train,y_train)
confusion_matrix(y_test,gnb.predict(X_test))

mnb = MultinomialNB()
mnb.fit(X_train,y_train)
confusion_matrix(y_test,mnb.predict(X_test))

bnb = BernoulliNB()
bnb.fit(X_train,y_train)
confusion_matrix(y_test,bnb.predict(X_test))

dt = DecisionTreeClassifier()
dt.fit(X_train,y_train)
confusion_matrix(y_test,dt.predict(X_test))

rf = RandomForestClassifier()
rf.fit(X_train,y_train)
confusion_matrix(y_test,rf.predict(X_test))

from sklearn.svm import SVC

svc = SVC(kernel = 'linear')
svc.fit(X_train,y_train)
confusion_matrix(y_test,svc.predict(X_test))

abc = AdaBoostClassifier(n_estimators = 50)
abc.fit(X_train,y_train)
confusion_matrix(y_test,abc.predict(X_test))

from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Dropout
from tensorflow.keras.utils import to_categorical
from tensorflow.keras.metrics import Recall, Precision

model = Sequential()
model.add(Dense(32,activation = 'relu'))
model.add(Dense(16, activation = 'relu'))
model.add(Dense(1,activation = 'sigmoid'))
model.compile(loss='binary_crossentropy',optimizer='adam',metrics = [Precision()])

model.fit(X_train,y_train,validation_data=(X_test, y_test), epochs = 50, batch_size = 36)

model.evaluate(X_test,y_test)

y_pred = model.predict(X_test) > 0.45

confusion_matrix(y_test,y_pred)
